pyproject.toml
[project]
name = "resume-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ollama>=0.4.8",
    "openai-agents>=0.0.15",
    "pypdf2>=3.0.1",
    "streamlit>=1.45.1",
]

